requirements.txt
ultralytics==8.4.21
opencv-python==4.13.0.92
PyYAML==6.0.2
requests==2.32.5
